# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PANOR Clinical Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      summary: List visits
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Visit list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitList"
    post:
      operationId: createVisit
      tags: [visits]
      summary: Start a new clinical visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVisitBody"
      responses:
        "201":
          description: Visit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}:
    get:
      operationId: getVisit
      tags: [visits]
      summary: Get visit details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
    patch:
      operationId: updateVisit
      tags: [visits]
      summary: Update visit notes and status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVisitBody"
      responses:
        "200":
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"

  # PRESCRIPTIONS
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      parameters:
        - name: visitId
          in: query
          schema:
            type: integer
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Prescriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionList"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create a prescription (triggers drug safety check)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionBody"
      responses:
        "201":
          description: Prescription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionResult"
  /prescriptions/drug-check:
    post:
      operationId: checkDrugInteraction
      tags: [prescriptions]
      summary: Check drug-drug interaction before prescribing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrugCheckBody"
      responses:
        "200":
          description: Drug interaction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrugCheckResult"

  # LAB ORDERS
...
components:
  schemas:
...
    Visit:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        doctorId:
          type: integer
        doctorName:
          type: string
        chiefComplaint:
          type: string
        clinicalNotes:
          type: string
          nullable: true
        aiSummary:
          type: string
          nullable: true
        diagnosis:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, cancelled]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, patientId, patientName, doctorId, doctorName, chiefComplaint, status, createdAt, updatedAt]
    VisitList:
      type: object
      properties:
        visits:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        total:
          type: integer
      required: [visits, total]
    CreateVisitBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        chiefComplaint:
          type: string
        symptoms:
          type: string
      required: [patientId, doctorId, chiefComplaint]
    UpdateVisitBody:
      type: object
      properties:
        clinicalNotes:
          type: string
        diagnosis:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
    Prescription:
      type: object
      properties:
        id:
          type: integer
        visitId:
          type: integer
        patientId:
          type: integer
        medication:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: string
          nullable: true
        safetyStatus:
          type: string
          enum: [safe, warning, blocked]
        interactionWarning:
          type: string
          nullable: true
        prescribedAt:
          type: string
      required: [id, visitId, patientId, medication, dosage, frequency, duration, safetyStatus, prescribedAt]
    PrescriptionList:
      type: object
      properties:
        prescriptions:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
      required: [prescriptions]
    PrescriptionResult:
      type: object
      properties:
        prescription:
          $ref: "#/components/schemas/Prescription"
        safetyCheck:
          $ref: "#/components/schemas/DrugCheckResult"
      required: [prescription, safetyCheck]
    CreatePrescriptionBody:
      type: object
      properties:
        visitId:
          type: integer
        patientId:
          type: integer
        medication:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        duration:
          type: string
        instructions:
          type: string
      required: [visitId, patientId, medication, dosage, frequency, duration]
    DrugCheckBody:
      type: object
      properties:
        patientId:
          type: integer
        newMedication:
          type: string
        currentMedications:
          type: array
          items:
            type: string
      required: [patientId, newMedication]
    DrugCheckResult:
      type: object
      properties:
        safe:
          type: boolean
        status:
          type: string
          enum: [safe, warning, blocked]
        interactions:
          type: array
          items:
            type: object
            properties:
              drug:
                type: string
              severity:
                type: string
              explanation:
                type: string
              suggestion:
                type: string
            required: [drug, severity, explanation]
        confidenceScore:
          type: number
        recommendation:
          type: string
      required: [safe, status, interactions, confidenceScore, recommendation]
...