# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PANOR Clinical Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { PrescriptionSafetyStatus } from "./prescriptionSafetyStatus";

export interface Prescription {
  id: number;
  visitId: number;
  patientId: number;
  medication: string;
  dosage: string;
  frequency: string;
  duration: string;
  instructions?: string | null;
  safetyStatus: PrescriptionSafetyStatus;
  interactionWarning?: string | null;
  prescribedAt: string;
}